feat(ui): Adding the admin dashboard
