+++ nexcart/apps/web/admin/src/Admin.css
@@ -1,4 +1,4 @@
-/* ============== APP SHELL ============== */
+
 .nx-admin {
     display: grid;
     grid-template-columns: 240px 1fr;
@@ -9,10 +9,9 @@
 .nx-admin__main {
     display: flex;
     flex-direction: column;
-    min-width: 0;            /* lets the table shrink instead of overflowing */
+    min-width: 0;
 }
 
-/* ============== SIDEBAR ============== */
 .nx-admin__sidebar {
     display: flex;
     flex-direction: column;
@@ -101,7 +100,7 @@
     color: var(--grey);
 }
 
-/* ============== TOOLBAR ============== */
+
 .nx-admin__toolbar {
     display: flex;
     align-items: center;
@@ -136,7 +135,7 @@
     background-color: var(--white);
 }
 
-/* ============== TABLE ============== */
+
 .nx-admin__table-card {
     margin: var(--size-20) var(--size-24);
     background-color: var(--white);
@@ -201,5 +200,67 @@
     height: var(--size-32);
     object-fit: cover;
     border-radius: var(--radius-6);
+    background-color: var(--off-white);
+}
+
+/* Navbar Styles */
+.nx-admin--navbar {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: var(--size-24);
+    padding: var(--size-16) var(--size-24);
+    background-color: var(--white);
+    border-bottom: 1px solid var(--light-grey);
+    flex-shrink: 0;
+}
+
+.nx-admin--navbar__left,
+.nx-admin--navbar__center,
+.nx-admin--navbar__right {
+    display: flex;
+    align-items: center;
+    gap: var(--size-16);
+}
+
+.nx-admin--navbar__center {
+    flex: 1;
+    justify-content: center;
+}
+
+.nx-admin--navbar__right {
+    gap: var(--size-12);
+}
+
+.nx-admin--header {
+    margin: 0;
+    font-size: 24px;
+    font-weight: var(--font-weight-600);
+    color: var(--grey);
+}
+
+.nx-admin--icon-btn {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: var(--size-40);
+    height: var(--size-40);
+    border: none;
+    background-color: transparent;
+    color: var(--grey);
+    cursor: pointer;
+    border-radius: var(--radius-8);
+    transition: background-color var(--transition-duration) ease;
+}
+
+.nx-admin--icon-btn:hover {
+    background-color: var(--off-white);
+}
+
+.nx-admin--avatar {
+    width: var(--size-40);
+    height: var(--size-40);
+    border-radius: var(--radius-100);
+    overflow: hidden;
     background-color: var(--off-white);
 }
